move unsafe code to non-appengine build files
